+++ src/writer.rs
@@ -1,12 +1,12 @@
-use futures_lite::io::AsyncWriteExt as _;
+use tokio::io::AsyncWriteExt as _;
 
 /// Writes ttyrec frames to a [`futures_lite::io::AsyncWrite`] instance.
-pub struct Writer<T: futures_lite::io::AsyncWrite> {
+pub struct Writer<T: tokio::io::AsyncWrite> {
     output: T,
     creator: crate::Creator,
 }
 
-impl<T: futures_lite::io::AsyncWrite + std::marker::Unpin + Send> Writer<T> {
+impl<T: tokio::io::AsyncWrite + std::marker::Unpin + Send> Writer<T> {
     /// Creates a new [`Writer`](Self) from a [`futures_lite::io::AsyncWrite`]
     /// instance.
     pub fn new(output: T) -> Self {